requirements.txt
# Required
google-genai>=1.0.0

# Optional: Chinese name → slug conversion
pypinyin>=0.48.0
